test: extract test subscriber to improve readability
